Fix block open file

+++ Block.py
@@ -1,19 +1,24 @@
-# from Record import *
+from Record import *
+import os
 
 
 class Block:
     def __init__(self, disk_name):
         """ Constructor """
         # open disk file
-        self.disk = open(disk_name, "w+")
+        if (os.path.isfile('./' + disk_name)):
+            self.disk = open(disk_name, "r+")
+        else:
+            self.disk = open(disk_name, "w+")
         # set size of block
         self.max_size = 8
         # set number of operations (read/write)
         self.n_op = 0
         # set list of records
         self.records = []
-        #record size
-        self.record_size=138
+        # record size
+        self.record_size = 138
+        self.pos = 0
 
     def __del__(self):
         """ Destructor """
@@ -40,6 +45,7 @@ class Block:
         if (len(self.records) >= self.max_size):
             self.persist(pos)
         self.records += [rec]
+        self.pos = pos
 
     def read(self, pos):
         """ Add records of pos position in disk file to this """
@@ -49,8 +55,8 @@ class Block:
         while (len(self.records) < self.max_size):
             self.records += [self.disk.read(self.record_size)]
 
-# r = Record("11111111111;54.037.661-5;estermoro@gmail.com;06/01/1952;Feminino;Yuri Matheus Antonia;5942.00")
-# b = Block("out.txt")
-# b.write(0, str(r))
-# b.persist(0)
+r = Record("11111111111;54.037.661-5;estermoro@gmail.com;06/01/1952;Feminino;Yuri Matheus Antonia;5942.00")
+b = Block("out.txt")
+b.write(500, str(r))
+# b.persist(500)
 # b.read(0)
